requirements.txt
aiofiles==23.2.1
aiogram==2.18
aiohttp==3.8.6
aiosignal==1.2.0
alembic==1.7.7
annotated-types==0.6.0
anyio==4.0.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.0.7
async-timeout==4.0.2
attrs==21.4.0
autopage==0.5.0
Babel==2.9.1
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.5
brotlipy==0.7.0
build==1.1.1
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.0.3
cliff==3.10.1
cmaes==0.8.2
cmd2==2.4.1
coloredlogs==15.0.1
colorlog==6.6.0
conda==4.13.0
conlleval==0.2
cycler==0.11.0
dataclasses-json==0.6.4
debugpy==1.6.2
decorator==5.1.1
deepdiff==6.7.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
dominate==2.6.0
emoji==2.10.1
entrypoints==0.4
exceptiongroup==1.1.3
executing==0.10.0
fastapi==0.110.0
fastjsonschema==2.16.1
filelock==3.13.1
filetype==1.2.0
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.10
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
flatbuffers==24.3.7
fonttools==4.33.3
frozenlist==1.3.1
fsspec==2024.3.1
gensim==4.0.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
graphlib_backport==1.1.0
grpcio==1.62.1
gspread==6.1.2
gunicorn==20.1.0
h11==0.14.0
h5py==3.7.0
httpcore==0.18.0
httptools==0.6.1
httpx==0.25.0
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==5.7.1
iniconfig==1.1.1
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
itsdangerous==2.0.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.7.2
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-http-over-ws==0.0.8
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
Keras==2.2.4
Keras-Applications==1.0.8
kiwisolver==1.4.2
kubernetes==29.0.0
langchain==0.1.13
langchain-community==0.0.29
langchain-core==0.1.33
langchain-openai==0.1.1
langchain-text-splitters==0.0.1
langdetect==1.0.9
langsmith==0.1.31
lxml==5.1.0
magic-filter==1.0.12
Mako==1.2.0
MarkupSafe==2.0.1
marshmallow==3.21.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.2
mypy-extensions==1.0.0
natsort==8.1.0
nbclient==0.6.6
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
nltk==3.7
notebook==6.4.12
oauthlib==3.2.2
onnxruntime==1.17.1
openai==0.28.0
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
optuna==2.10.0
ordered-set==4.1.0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pbr==5.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.1
pluggy==1.0.0
posthog==3.5.0
prettytable==3.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==4.25.3
psutil==5.9.1
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycosat==0.6.3
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.12.0
pyparsing==3.0.9
pypdf==4.1.0
pyperclip==1.8.2
PyPika==0.48.9
pyproject_hooks==1.0.0
pyrsistent==0.18.1
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-iso639==2024.2.7
python-magic==0.4.27
python-telegram-bot==20.6
python2==1.2
pytz==2022.1
PyYAML==6.0
pyzmq==23.2.0
qtconsole==5.3.1
QtPy==2.2.0
rapidfuzz==3.7.0
ratelimit==2.2.1
regex==2022.1.18
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.1.1
Send2Trash==1.8.0
sentry-sdk==1.8.0
sklearn==0.0
smart-open==5.2.1
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==2.0.29
SQLAlchemy-Utils==0.38.3
stack-data==0.4.0
starlette==0.36.3
stevedore==3.5.0
sympy==1.12
tabulate==0.9.0
telegram==0.0.1
tenacity==8.2.3
termcolor==1.1.0
terminado==0.15.0
threadpoolctl==3.1.0
tiktoken==0.6.0
tinycss2==1.1.1
tokenizers==0.15.2
tomli==2.0.1
torch==1.12.0
torchvision==0.13.0
tornado==6.2
tqdm==4.66.2
traitlets==5.3.0
typer==0.10.0
typing-inspect==0.9.0
typing_extensions==4.10.0
unstructured==0.11.8
unstructured-client==0.22.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
visitor==0.1.3
waitress==2.1.2
watchfiles==0.21.0
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==2.0.3
widgetsnbextension==3.6.1
wrapt==1.16.0
WTForms==3.0.1
yarl==1.8.1
zipp==3.8.0